use-cases/03-call-center-analytics/requirements.txt
# Direct dependencies only

python-dotenv>=1.2.1
speechmatics-batch>=0.4.4
